requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Telegram bot
python-telegram-bot[webhooks]>=21.3

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1

# Task queue
celery>=5.4.0
redis>=5.0.4

# AI / LLM
anthropic>=0.30.1

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.5

# Data validation
pydantic>=2.7.1
pydantic-settings>=2.2.1

# Utilities
python-dotenv>=1.0.1
pytz>=2024.1
python-dateutil>=2.9.0

# Logging
structlog>=24.1.0
